refactor(architecture): promote DEFAULT_REGISTRY_PATTERNS, single-source POOL_SERIAL_THRESHOLD, extract PROJECT_ROOT leaf (refs #3993)
