+++ .agents/scripts/evidence-gate.js
@@ -47,7 +47,7 @@
 import { spawnSync } from 'node:child_process';
 import { parseArgs } from 'node:util';
 import { runAsCli } from './lib/cli-utils.js';
-import { PROJECT_ROOT } from './lib/config-resolver.js';
+import { PROJECT_ROOT } from './lib/project-root.js';
 import { gitSpawn } from './lib/git-utils.js';
 import { Logger } from './lib/Logger.js';
 import {
