+++ .agents/scripts/git-rebase-and-resolve.js
@@ -40,7 +40,7 @@
 
 import { parseArgs } from 'node:util';
 import { runAsCli } from './lib/cli-utils.js';
-import { PROJECT_ROOT } from './lib/config-resolver.js';
+import { PROJECT_ROOT } from './lib/project-root.js';
 import { gitSpawn } from './lib/git-utils.js';
 import { Logger } from './lib/Logger.js';
